Refactor rename View -> CalculatorView, Validator -> ExpressionValidator
