+++ src/main/java/calculator/Main.java
@@ -4,11 +4,11 @@ public class Main {
     public static void main(String[] args) {
         int result;
 
-        String expression = View.input();
-        if (Validator.isValid(expression)) {
+        String expression = CalculatorView.input();
+        if (ExpressionValidator.isValid(expression)) {
             StringCalculator stringCalculator = new StringCalculator(expression);
             result = stringCalculator.calculate();
-            View.output(result);
+            CalculatorView.output(result);
         }
     }
 }
