+++ src/main/java/calculator/StringCalculator.java
@@ -4,11 +4,6 @@ import java.util.LinkedList;
 import java.util.Queue;
 
 public class StringCalculator {
-    private final static String PLUS = "+";
-    private final static String MINUS = "-";
-    private final static String MULTIPLE = "*";
-    private final static String DIVIDE = "/";
-
     private Queue<Integer> numbers = new LinkedList<>();
     private Queue<String> operators = new LinkedList<>();
 
@@ -18,13 +13,11 @@ public class StringCalculator {
     }
 
     private void divideExpression(Queue<Integer> numbers, Queue<String> operators, String[] expression) {
-        for (int i = 0; i < expression.length; i++) {
-            if (i % 2 == 0) { // even
-                numbers.add(Integer.parseInt(expression[i]));
-            }
-            if (i % 2 == 1) { // odd
-                operators.add(expression[i]);
-            }
+        for(int i = 0; i < expression.length; i = i + 2){
+            numbers.add(Integer.parseInt(expression[i]));
+        }
+        for(int i = 1; i < expression.length; i = i + 2){
+            operators.add(expression[i]);
         }
     }
